% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_megan.R
\name{theme_megan}
\alias{theme_megan}
\alias{megan_theme}
\title{Megan's ggplot2 theme}
\usage{
theme_megan(base_size = 11, base_family = "", ...)

megan_theme(base_size = 11, base_family = "", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2]{theme}}}
}
\value{
An object as returned by \code{\link[ggplot2]{theme}}
}
\description{
Megan's ggplot2 theme: black border, no ticks, subtile grid lines.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(y=mpg, x=disp, colour=factor(cyl))) +
    geom_point() + theme_megan()

}
\seealso{
\code{\link[ggplot2]{theme}}
}
